backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-can==4.4.2
pydantic==2.9.2
